+++ bssopenapi-20230930/src/Models/DescribeCouponRequest.php
@@ -55,9 +55,19 @@ class DescribeCouponRequest extends Model
     public $expireStartDate;
 
     /**
+     * @var int
+     */
+    public $maxResults;
+
+    /**
      * @var string
      */
     public $nbid;
+
+    /**
+     * @var string
+     */
+    public $nextToken;
 
     /**
      * @var int
@@ -78,7 +88,9 @@ class DescribeCouponRequest extends Model
         'effectiveStartTime' => 'EffectiveStartTime',
         'expireEndDate' => 'ExpireEndDate',
         'expireStartDate' => 'ExpireStartDate',
+        'maxResults' => 'MaxResults',
         'nbid' => 'Nbid',
+        'nextToken' => 'NextToken',
         'pageSize' => 'PageSize',
         'status' => 'Status',
     ];
@@ -137,8 +149,16 @@ class DescribeCouponRequest extends Model
             $res['ExpireStartDate'] = $this->expireStartDate;
         }
 
+        if (null !== $this->maxResults) {
+            $res['MaxResults'] = $this->maxResults;
+        }
+
         if (null !== $this->nbid) {
             $res['Nbid'] = $this->nbid;
+        }
+
+        if (null !== $this->nextToken) {
+            $res['NextToken'] = $this->nextToken;
         }
 
         if (null !== $this->pageSize) {
@@ -203,8 +223,16 @@ class DescribeCouponRequest extends Model
             $model->expireStartDate = $map['ExpireStartDate'];
         }
 
+        if (isset($map['MaxResults'])) {
+            $model->maxResults = $map['MaxResults'];
+        }
+
         if (isset($map['Nbid'])) {
             $model->nbid = $map['Nbid'];
+        }
+
+        if (isset($map['NextToken'])) {
+            $model->nextToken = $map['NextToken'];
         }
 
         if (isset($map['PageSize'])) {